+++ source/language/en.js
@@ -626,7 +626,8 @@ Hot Keys:
       prompt: {
         btns: {
           ok: 'Update',
-          no: 'Cancel'
+          no: 'Cancel',
+          changelog: 'ChangeLog'
         },
         body: (ver) => `Found new version v${ver}, update now?`,
         loader_body: (ver) => `The new version of the Loader v${ver} has been released, the currently used loader will not be available, will you quit the program and go to download the latest version of the Loader?`,